+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - Indicate progress while downloading
+- Lock current dirs to actual OS/ARCH
 
 ## [0.4.0] - 2018-03-03
 ### Added